Analyzing Life Satisfaction and GDP Relationship: Polynomial Regression and K-Nearest neighbours.
Working with the data of the data frame of Life satisfaction and the GDP per capita (full_country_stats):
Utilize polynomial regression with a degree of 8 and visualize the outcomes. Determine the life satisfaction index (LSI) for a nation with a GDP of 97000. Employ the three closest neighboring countries to the one with a GDP of 97000 to approximate its LSI.
Compare the forecasts generated by the two methods and provide insights. Generate predictions using polynomial degrees ranging from 1 to 10.
Analyze the findings.